Huh... never committed this :/
last week: finished most of the major parts!

color tools were fixed to spread out colors into an array

wheel piece was tweaked a little bit

this week:
fixed the point-tick detection to actually function right :P

created 3/4ths of the tiny triangle on the top of the wheel. It's getting late tonight so it will have to wait.