I apparently take a while to commit things :/
Attempted to make fade ins / fade outs.  It didn't work. Further inspection revealed it was because I was 1. doing it backwards, and 2. The method was designed in a way  not friendly to fades

next attempt will be to potentiallly re-write the animate script. (We'll see if it goes to that)

Started working on text in wheel pieces